chore: no react.lazy
